refactor: move getEntitiesWithEntityTypeProvided to OlsEntityAPI and move EntityListWidgetProps to types.ts
