+++ packages/react/src/components/widgets/EntityListWidget/EntityListWidgetStories.ts
@@ -1,6 +1,5 @@
 // EntityListWidgetStories.ts
 
-
 import * as globals from "../../../app/globals";
 import { EBI_API_ENDPOINT } from "../../../app/globals";
 import {
